+++ jdbc/src/main/java/nextstep/jdbc/JdbcTemplate.java
@@ -5,6 +5,10 @@ import java.util.List;
 
 import javax.sql.DataSource;
 
+import nextstep.jdbc.callback.ExecuteCallback;
+import nextstep.jdbc.callback.RowMapperCallback;
+import nextstep.jdbc.callback.StatementCallback;
+
 public class JdbcTemplate {
 
 	private final JdbcUtil jdbcUtil;
