+++ jdbc/src/main/java/nextstep/jdbc/callback/RowMapperCallback.java
@@ -1,4 +1,4 @@
-package nextstep.jdbc;
+package nextstep.jdbc.callback;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
